Rename hand-detection package to hand-pose-detection
