+++ hand-pose-detection/.npmignore
