+++ hand-pose-detection/demos/live_video/src/camera.js
