+++ hand-pose-detection/demos/live_video/src/option_panel.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  * =============================================================================
  */
-import * as handdetection from '@tensorflow-models/hand-detection';
+import * as handdetection from '@tensorflow-models/hand-pose-detection';
 import * as tf from '@tensorflow/tfjs-core';
 
 import * as params from './params';
