+++ hand-pose-detection/demos/live_video/src/stats_panel.js
