+++ hand-pose-detection/demos/live_video/src/util.js
