+++ hand-pose-detection/karma.conf.js
