+++ hand-pose-detection/src/constants.ts
