+++ hand-pose-detection/src/create_detector.ts
