+++ hand-pose-detection/src/hand_detector.ts
