+++ hand-pose-detection/src/mediapipe/constants.ts
