+++ hand-pose-detection/src/mediapipe/detector.ts
