+++ hand-pose-detection/src/mediapipe/detector_utils.ts
