+++ hand-pose-detection/src/mediapipe/mediapipe_test.ts
@@ -15,18 +15,23 @@
  * =============================================================================
  */
 
-import * as tf from '@tensorflow/tfjs-core';
 // tslint:disable-next-line: no-imports-from-dist
-import {ALL_ENVS, BROWSER_ENVS, describeWithFlags} from '@tensorflow/tfjs-core/dist/jasmine_util';
+import {BROWSER_ENVS, describeWithFlags} from '@tensorflow/tfjs-core/dist/jasmine_util';
 // tslint:disable-next-line: no-imports-from-dist
 import {expectArraysClose} from '@tensorflow/tfjs-core/dist/test_util';
 
-import * as handDetection from '../index';
+import * as handPoseDetection from '../index';
 import {loadImage} from '../shared/test_util';
 
-// Measured in pixels.
-const EPSILON_IMAGE = 35;
+import {MediaPipeHandsMediaPipeModelConfig} from './types';
 
+const MEDIAPIPE_MODEL_CONFIG: MediaPipeHandsMediaPipeModelConfig = {
+  runtime: 'mediapipe',
+  solutionPath: 'base/node_modules/@mediapipe/hands'
+};
+
+// In pixels.
+const EPSILON_IMAGE = 20;
 // ref:
 // https://github.com/google/mediapipe/blob/master/mediapipe/python/solutions/hands_test.py
 const EXPECTED_HAND_KEYPOINTS_PREDICTION = [
@@ -44,54 +49,8 @@ const EXPECTED_HAND_KEYPOINTS_PREDICTION = [
   ]
 ];
 
-describeWithFlags('MediaPipeHands', ALL_ENVS, () => {
-  let timeout: number;
-
-  beforeAll(() => {
-    timeout = jasmine.DEFAULT_TIMEOUT_INTERVAL;
-    jasmine.DEFAULT_TIMEOUT_INTERVAL = 120000;  // 2mins
-  });
-
-  afterAll(() => {
-    jasmine.DEFAULT_TIMEOUT_INTERVAL = timeout;
-  });
-
-  it('estimateHands does not leak memory.', async () => {
-    const startTensors = tf.memory().numTensors;
-
-    // Note: this makes a network request for model assets.
-    const detector = await handDetection.createDetector(
-        handDetection.SupportedModels.MediaPipeHands, {runtime: 'tfjs'});
-    const input: tf.Tensor3D = tf.zeros([128, 128, 3]);
-
-    const beforeTensors = tf.memory().numTensors;
-
-    await detector.estimateHands(input);
-
-    expect(tf.memory().numTensors).toEqual(beforeTensors);
-
-    detector.dispose();
-    input.dispose();
-
-    expect(tf.memory().numTensors).toEqual(startTensors);
-  });
-
-  it('throws error when runtime is not set.', async (done) => {
-    try {
-      await handDetection.createDetector(
-          handDetection.SupportedModels.MediaPipeHands);
-      done.fail('Loading without runtime succeeded unexpectedly.');
-    } catch (e) {
-      expect(e.message).toEqual(
-          `Expect modelConfig.runtime to be either ` +
-          `'tfjs' or 'mediapipe', but got undefined`);
-      done();
-    }
-  });
-});
-
-describeWithFlags('MediaPipeHands static image ', BROWSER_ENVS, () => {
-  let detector: handDetection.HandDetector;
+describeWithFlags('MediaPipe Hands multi hands ', BROWSER_ENVS, () => {
+  let detector: handPoseDetection.HandDetector;
   let image: HTMLImageElement;
   let timeout: number;
 
@@ -106,27 +65,17 @@ describeWithFlags('MediaPipeHands static image ', BROWSER_ENVS, () => {
   });
 
   it('test.', async () => {
-    const startTensors = tf.memory().numTensors;
-
     // Note: this makes a network request for model assets.
-    detector = await handDetection.createDetector(
-        handDetection.SupportedModels.MediaPipeHands, {runtime: 'tfjs'});
+    const model = handPoseDetection.SupportedModels.MediaPipeHands;
+    detector =
+        await handPoseDetection.createDetector(model, MEDIAPIPE_MODEL_CONFIG);
 
-    const beforeTensors = tf.memory().numTensors;
-
-    const result = await detector.estimateHands(
-        image, {staticImageMode:
-                    true} as handDetection.MediaPipeHandsTfjsEstimationConfig);
+    const result = await detector.estimateHands(image, {staticImageMode: true});
     const keypoints = result.map(
         hand => hand.keypoints.map(keypoint => [keypoint.x, keypoint.y]));
 
     expectArraysClose(
         keypoints, EXPECTED_HAND_KEYPOINTS_PREDICTION, EPSILON_IMAGE);
-
-    expect(tf.memory().numTensors).toEqual(beforeTensors);
-
     detector.dispose();
-
-    expect(tf.memory().numTensors).toEqual(startTensors);
   });
 });
