+++ hand-pose-detection/src/mediapipe/types.ts
