+++ hand-pose-detection/src/setup_test.ts
