+++ hand-pose-detection/src/shared
