+++ hand-pose-detection/src/tfjs/calculators/hand_landmarks_to_rect.ts
