+++ hand-pose-detection/src/tfjs/constants.ts
