+++ hand-pose-detection/src/tfjs/detector.ts
