+++ hand-pose-detection/src/tfjs/detector_utils.ts
