+++ hand-pose-detection/src/tfjs/types.ts
