+++ hand-pose-detection/src/types.ts
