+++ hand-pose-detection/test_data/hands.jpg
