+++ hand-pose-detection/tsconfig.json
