+++ hand-pose-detection/tsconfig.test.json
